Remove the isDate() validator
It attempts to validate too many different formats, and as a
result there are edge cases which cannot be resolved.

Additionally, since we make use of the underlying JS engine's date
parser, there have been cases where a date might be valid in one
engine but not valid in another (e.g. #602).

If you need date validation, try using isISO8601(), or try rolling
your own regex which accepts dates in a specific format, e.g.

   validator.matches(date, /^\d{2}/\d{2}/\d{4}$/)

The validator might make a return at some point, but if so it would
require a format string (e.g. #561).