fix: extra validation for dates
This commit adds an additional test for dates
over and above passing the pattern match.

It checks to assertain that the date is a
valid date. For examples, dates like
2009-02-29 are caught as invalid.

Additionally, it also fixes a minor bug
with the `weekDate` that was wrongfully
leaving out W53 as an invalid date.

fixes #772