refactor(isTaxID): Prepare to support more sanitization regexes Different locales might have specific needs wrt acceptable symbols in TINs- in some all are omitted during validation, while others only allow to skip a subset.
A new object `sanitizeRegexes` has replaced the previous array, where
locale-specific skippable symbol classes are to be placed. When all
symbols can be omitted the new variable `allsymbols` is referenced.

Aliases have also been added for the new object in line with the
others and the isTaxID function checks for the locale's inclusion in
`sanitizeRegexes`.