chore: add rik and rubin as maintainers
I would like to propose to add both Rik and Rubiin as maintainers
on the project because of their dedication to the project in the
past 2+ years. They have been serving as triagers / moderators
in the past period.

I will also look forward to hand over the keys to a new guard
very soon.

Long live validator.js <3