Add Norwegian passport validation
The regex is constructed based on the Norwegian Police's control
guide:
https://www.politiet.no/globalassets/tjenester-admin/pass-og-id-kort/control-guide-for-norwegian-passports.pdf
(ISBN 978-82-8256-097-9).

It matches Norwegian passport numbers starting with C, K, or FG, FH,
FJ, FL, and FK prefixes, followed by alphanumeric characters.

The guide is not very detailed. It does not specify passport number
length, so this is not validated.